Мини рефакторинг
