+++ src/const.js
@@ -3,3 +3,27 @@ export const FilterType = {
   FUTURE: `future`,
   PAST: `past`
 };
+
+export const SortType = {
+  EVENT: `event`,
+  TIME: `time`,
+  PRICE: `price`,
+};
+
+export const Mode = {
+  DEFAULT: `default`,
+  EDIT: `edit`,
+};
+
+export const EventType = {
+  TAXI: `taxi`,
+  BUS: `bus`,
+  TRAIN: `train`,
+  SHIP: `ship`,
+  TRANSPORT: `transport`,
+  DRIVE: `drive`,
+  FLIGHT: `flight`,
+  CHECK_IN: `check-in`,
+  SIGHTSEEING: `sightseeing`,
+  RESTAURANT: `restaurant`,
+};
