+++ src/controllers/trip.js
@@ -1,10 +1,11 @@
-import TripSortComponent, {SortType} from "../components/trip-sort.js";
+import TripSortComponent from "../components/trip-sort.js";
 import NewEventComponent from "../components/new-event.js";
 import DayListComponent from "../components/day-list.js";
 import DayInfoComponent from "../components/day-info.js";
 import NoEventsComponent from "../components/no-events.js";
 import TripInfoComponent from "../components/trip-info.js";
 import PointController from "./point.js";
+import {SortType} from "../const.js";
 import {render, RenderPosition, remove} from "../utils/render.js";
 
 const SHOWING_EVENTS_COUNT = 20;
